rename typo in file
