+++ packages/react/src/experimental/OneDataCollection/CollectionActions/index.ts
@@ -1,1 +1,1 @@
-export * from "./ColletionActions"
+export * from "./CollectionActions"
