+++ packages/react/src/experimental/OneDataCollection/index.tsx
@@ -6,7 +6,7 @@ import { Icon } from "../../components/Utilities/Icon"
 import { Spinner } from "../../icons/app"
 import { OneActionBar } from "../OneActionBar"
 import { useLayout } from "../PageLayouts/LayoutProvider"
-import { CollectionActions } from "./CollectionActions/ColletionActions"
+import { CollectionActions } from "./CollectionActions/CollectionActions"
 import { FiltersChipsList } from "./Filters/Components/FiltersChipsList"
 import { FiltersControls } from "./Filters/Components/FiltersControls"
 import type { FiltersDefinition, FiltersState } from "./Filters/types"
